+++ package/telegram_client/lib/isolate/isolate.dart
@@ -1,1 +1,1 @@
-export "isolate_io.dart";
+export "isolate_io.dart" if (dart.library.html) "isolate_web.dart";