+++ package/telegram_client/lib/scheme/scheme.dart
@@ -1,1 +1,2 @@
 export "tdlib_client.dart";
+export "mtproto_client.dart";