+++ package/telegram_client/lib/telegram_client.dart
@@ -5,3 +5,4 @@ export "package:telegram_client/telegram_bot_api/extension/telegram_bot_api.dart
 export "package:telegram_client/extensions/extensions.dart";
 export "package:telegram_client/util/util.dart";
 export "package:telegram_client/scheme/scheme.dart";
+export "package:telegram_client/mtproto/mtproto.dart";