+++ package/telegram_client/template/telegram_userbot_mtproto_template/bin/telegram_userbot_mtproto_template.dart
@@ -1,0 +1,6 @@
+import 'package:telegram_userbot_mtproto_template/telegram_userbot_mtproto_template.dart'
+    as telegram_userbot_mtproto_template;
+
+void main(List<String> arguments) {
+  print('Hello world: ${telegram_userbot_mtproto_template.calculate()}!');
+}