+++ package/telegram_client/template/telegram_userbot_mtproto_template/test/telegram_userbot_mtproto_template_test.dart
@@ -1,0 +1,8 @@
+import 'package:telegram_userbot_mtproto_template/telegram_userbot_mtproto_template.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('calculate', () {
+    expect(calculate(), 42);
+  });
+}